+++ extension/test/unit/user/auth-strategy/basic-test.js
@@ -2,7 +2,7 @@ const assert = require('assert')
 const expect = require('chai').expect
 const step = require('../../../../user/auth-strategy/basic')
 
-describe('login', () => {
+describe('basic login', () => {
   let request = null
 
   const context = {
@@ -106,13 +106,9 @@ describe('login', () => {
 
   it('should return an error because deleting guest tokens failed', (done) => {
     const magentoResponse = {
-      success: [
-        {
-          'expires_in': 3600,
-          'access_token': 'a1',
-          'refresh_token': 'r1'
-        }
-      ]
+      'expires_in': 3600,
+      'access_token': 'a1',
+      'refresh_token': 'r1'
     }
 
     request.post = (options, cb) => {